+++ native/ivory/src/ivory_bone.cpp
@@ -32,7 +32,7 @@ void IvoryBone::_bind_methods() {
 
 void IvoryBone::set_skeleton(const PackedVector2Array &head,
 		const PackedVector2Array &tail, const PackedInt32Array &parent) {
-	const int n = std::min(head.size(), tail.size());
+	const int n = std::min((int)head.size(), (int)tail.size());
 	head_.resize(n);
 	tail_.resize(n);
 	parent_.resize(n);
